changed way of extracting addHouseForm component
